Fix appending backup a zillion times

+++ MSET9_installer_script/mset9.py
@@ -272,10 +272,13 @@ def injection():
 	
 	if os.path.exists(realId1Path):
 		prinfo("Backing up real Id1...")
-		os.rename(realId1Path, realId1Path + realId1BackupTag)
-	
-	id1 += realId1BackupTag
-	realId1Path = f"{id0}/{id1}"
+		if not os.path.exists(realId1Path + realId1BackupTag):
+			os.rename(realId1Path, realId1Path + realId1BackupTag)
+			id1 += realId1BackupTag
+			realId1Path = f"{id0}/{id1}"
+		else:
+			prinfo("Skipping backup because a backup already exists!")
+
 
 	prgood("MSET9 successfully injected!")
